feat: add a function, that checks existence of a film in the database, function added to all interactions wiche based on adding or changing films in database.

+++ data_base/sqlite_db.py
@@ -23,32 +23,60 @@ def sql_start():
     base.commit()
 
 
-async def sql_add_film(film: str):
+def sql_check_film(film: str) -> tuple:
     """
-    Adds a film to the films table in the database.
+    Check if a film exists in the database.
+    :param film: The name of the film to check.
+    :return: A tuple containing a single element indicating the existence of the film.
+               If the film exists, the tuple contains the film name. If the film doesn't
+               exist, the tuple is empty.
+    """
+    cur.execute("SELECT film FROM films WHERE film = ?", (film,))
+    film_existence = cur.fetchone()
+    return film_existence
+
+
+async def sql_add_film(film: str) -> str:
+    """
+    Adds a film to the films table in the database if this is not already exists there.
 
     This asynchronous function inserts a new film entry into the films table
     with the provided film name and status. It then commits the changes to the database.
 
     :param film: A string containing the film name and its status.
-    :return: None
+    :return: A message indicating the result of the operation.
     """
-    cur.execute("INSERT INTO films VALUES (?, 'active')", (film,))
-    base.commit()
+    # Check if the film exists in a database
+    film_existence = sql_check_film(film)
+
+    # Based on the film existence, adds it to the database or notify the user, that the film is already in the database
+    if film_existence == (film,):
+        return f"'{film}' already exist in the database."
+    else:
+        cur.execute("INSERT INTO films VALUES (?, 'active')", (film,))
+        base.commit()
+        return f"Added '{film}' to the database."
 
 
-async def sql_change_status(film: str):
+async def sql_change_status(film: str) -> str:
     """
-    Changes the status of a film to 'watched' in the films table.
+    Changes the status of a film to 'watched' in the films table, if the film exists there.
 
-    This asynchronous function updates the status of a film in the films table
+    This function updates the status of a film in the films table
     to 'watched' based on the provided film name. It then commits the changes to the database.
 
     :param film: A string containing the film name.
-    :return: None
+    :return: A message indicating the result of the operation.
     """
-    cur.execute("UPDATE films SET status = 'watched' WHERE film = ?", (film,))
-    base.commit()
+
+    # Check if the film exists in a database
+    film_existence = sql_check_film(film)
+    if film_existence == (film,):
+        cur.execute("UPDATE films SET status = 'watched' WHERE film = ?", (film,))
+        base.commit()
+        return f"Status of '{film}' changed to 'watched'."
+    else:
+        return f"There is no '{film}' in the database. Please check for proper name of the film."
 
 
 async def sql_show_suggestions(message, status):
@@ -68,18 +96,25 @@ async def sql_show_suggestions(message, status):
         await bot.send_message(message.from_user.id, text=film[0])
 
 
-async def sql_delete_film(film: str):
+async def sql_delete_film(film: str) -> str:
     """
-    Deletes a film from the films table in the database.
+    Deletes a film from the films table in the database if it exists there.
 
-    This asynchronous function removes a film entry from the films table based on the provided film name.
+    This function removes a film entry from the films table based on the provided film name.
     It then commits the changes to the database.
 
     :param film: A str containing the film name.
     :return: None
     """
-    cur.execute("DELETE FROM films WHERE film == ?", (film,))
-    base.commit()
+    print("delete")
+    film_existence = sql_check_film(film)
+    print(film_existence)
+    if film_existence == (film,):
+        cur.execute("DELETE FROM films WHERE film == ?", (film,))
+        base.commit()
+        return f"Film '{film}' deleted."
+    else:
+        return f"There was not '{film}' in database. Please check for proper name of the film. "
 
 
 async def sql_fetch_random(message, num_films):
